refactor csv reading into writer files
