+++ src/main/java/me/aarondmello/driver/DataReader.java
@@ -7,4 +7,5 @@ import java.io.IOException;
 public interface DataReader{
     void readFromStarterFile(BufferedReader reader, Tournament tournament) throws IOException;
     Tournament readFromInProgressFile(BufferedReader reader);
+    void readRoundResults(Tournament tournament, String fileName);
 }
